requirements.txt
# torch~=2.1.0
torchmetrics~=0.6.0
# linear-operator~=0.1.1
submitit~=1.5.1     # parallel execution helper
python-dotenv       # loading env variables from .env file
matplotlib~=3.7.0   # plotting
seaborn             # plotting
wheel               # Eases installation of some deps
einops~=0.6.0       # Eases operations on tensors
graphviz~=0.20.0    # Plotting graphical models
ipykernel           # Run this in jupyter
jupyter             # Run this in jupyter
jupyterlab          # Run this in jupyter
dash~=2.5.0
jupyter-dash~=0.4.2

## debugging affordances
lovely-tensors
ipdb~=0.13.13
memory-profiler>=0.61.0
tensorboard~=2.12.0

neuraloperator~=0.3.0
phiflow~=2.5.1
# scalene~=1.5.31.1
tqdm~=4.65.0
tueplots~=0.0.16
# annoying unused, undeclared dependencies  from neuraloperator
torch-harmonics~=0.7.1
wandb~=0.17.5
